Riacarmin/ch59/user should be able to edit a post

+++ ng/src/app/containers/view-photo/view-photo.component.ts
@@ -16,7 +16,7 @@ import { UsersService } from '../../services/users.service';
 export class ViewPhotoComponent implements OnInit {
   post: Post;
   self: Boolean;
-  contentChecked: Boolean;
+  editing: Boolean;
   constructor(
     private postsService: PostsService,
     private route: ActivatedRoute,
@@ -25,6 +25,7 @@ export class ViewPhotoComponent implements OnInit {
   ) { }
 
   ngOnInit(): void {
+    this.editing = false;
     this.route.paramMap
       .switchMap((params: ParamMap) => this.postsService.getPostById(params.get('id')))
       .subscribe(post => {
@@ -34,9 +35,20 @@ export class ViewPhotoComponent implements OnInit {
   }
 
   onEdit() {
-    alert('NOT READY!');
+    this.editing = !this.editing;
   }
   onDelete() {
-    this.postsService.deletePostById(this.post._id, this.post.author.userId);
+    if (window.confirm('Do you really want to delete your post?')) {
+      this.postsService.deletePostById(this.post._id, this.post.author.userId);
+    }
+  }
+
+  onUpdatePost() {
+    if (self) {
+      this.postsService.updatePost(this.post);
+      this.editing = false;
+    } else {
+      alert('Something is not right...');
+    }
   }
 }
